requirements.txt
autogen-agentchat
autogen-ext[openai]
autogen-ext[azure]
autogen-ext[magentic-one]
python-dotenv
langchain-community
agentops
duckduckgo-search
